Split out individual component parsers
