+++ src/layout/parser.rs
@@ -1,22 +1,35 @@
 //! Provides the parser for layout files of the original LiveSplit.
 
 use super::{Component, Layout};
-use crate::component::{
-    blank_space, current_comparison, current_pace, delta, detailed_timer, graph,
-    possible_time_save, previous_segment, separator, splits, sum_of_best, text, timer, title,
-    total_playtime,
-};
-use crate::settings::{Alignment, Color, Gradient, ListGradient};
-use crate::timing::{
-    formatter::{Accuracy, DigitsFormat},
-    TimingMethod,
-};
-use crate::xml_util::{
-    end_tag, parse_base, parse_children, text, text_as_escaped_bytes_err, text_err, text_parsed,
-    Error as XmlError, Tag,
+use crate::{
+    component::separator,
+    settings::{Alignment, Color, Gradient, ListGradient},
+    timing::{
+        formatter::{Accuracy, DigitsFormat},
+        TimingMethod,
+    },
+    xml_util::{
+        end_tag, parse_base, parse_children, text, text_as_escaped_bytes_err, text_err,
+        text_parsed, Error as XmlError, Tag,
+    },
 };
 use quick_xml::Reader;
 use std::io::BufRead;
+
+mod blank_space;
+mod current_comparison;
+mod current_pace;
+mod delta;
+mod detailed_timer;
+mod graph;
+mod possible_time_save;
+mod previous_segment;
+mod splits;
+mod sum_of_best;
+mod text;
+mod timer;
+mod title;
+mod total_playtime;
 
 quick_error! {
     /// The Error type for parsing layout files of the original LiveSplit.
@@ -316,921 +329,6 @@ where
     })
 }
 
-fn blank_space_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut blank_space::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-    let mut background_builder = GradientBuilder::new();
-
-    parse_children::<_, _, Error>(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if tag.name() == b"SpaceHeight" {
-                text_parsed(reader, tag.into_buf(), |h| settings.height = h)
-            } else {
-                // FIXME:
-                // SpaceWidth
-                end_tag(reader, tag.into_buf())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    settings.background = background_builder.build();
-
-    Ok(())
-}
-
-fn current_comparison_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut current_comparison::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-    let mut background_builder = GradientBuilder::new();
-    let (mut override_label, mut override_value) = (false, false);
-
-    parse_children(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if tag.name() == b"TextColor" {
-                color(reader, tag.into_buf(), |c| settings.label_color = Some(c))
-            } else if tag.name() == b"OverrideTextColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_label = b)
-            } else if tag.name() == b"TimeColor" {
-                color(reader, tag.into_buf(), |c| settings.value_color = Some(c))
-            } else if tag.name() == b"OverrideTimeColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_value = b)
-            } else if tag.name() == b"Display2Rows" {
-                parse_bool(reader, tag.into_buf(), |b| settings.display_two_rows = b)
-            } else {
-                // FIXME:
-                // Font1
-                // Font2
-                // OverrideFont1
-                // OverrideFont2
-                end_tag(reader, tag.into_buf())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    if !override_label {
-        settings.label_color = None;
-    }
-    if !override_value {
-        settings.value_color = None;
-    }
-    settings.background = background_builder.build();
-
-    Ok(())
-}
-
-fn current_pace_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut current_pace::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-    let mut background_builder = GradientBuilder::new();
-    let (mut override_label, mut override_value) = (false, false);
-
-    parse_children(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if tag.name() == b"TextColor" {
-                color(reader, tag.into_buf(), |c| settings.label_color = Some(c))
-            } else if tag.name() == b"OverrideTextColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_label = b)
-            } else if tag.name() == b"TimeColor" {
-                color(reader, tag.into_buf(), |c| settings.value_color = Some(c))
-            } else if tag.name() == b"OverrideTimeColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_value = b)
-            } else if tag.name() == b"Accuracy" {
-                accuracy(reader, tag.into_buf(), |a| settings.accuracy = a)
-            } else if tag.name() == b"Comparison" {
-                comparison_override(reader, tag.into_buf(), |v| settings.comparison_override = v)
-            } else if tag.name() == b"Display2Rows" {
-                parse_bool(reader, tag.into_buf(), |b| settings.display_two_rows = b)
-            } else {
-                end_tag(reader, tag.into_buf())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    if !override_label {
-        settings.label_color = None;
-    }
-    if !override_value {
-        settings.value_color = None;
-    }
-    settings.background = background_builder.build();
-
-    Ok(())
-}
-
-fn delta_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut delta::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-    let mut background_builder = GradientBuilder::new();
-    let mut override_label = false;
-
-    parse_children(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if tag.name() == b"TextColor" {
-                color(reader, tag.into_buf(), |c| settings.label_color = Some(c))
-            } else if tag.name() == b"OverrideTextColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_label = b)
-            } else if tag.name() == b"Accuracy" {
-                accuracy(reader, tag.into_buf(), |a| settings.accuracy = a)
-            } else if tag.name() == b"Comparison" {
-                comparison_override(reader, tag.into_buf(), |v| settings.comparison_override = v)
-            } else if tag.name() == b"Display2Rows" {
-                parse_bool(reader, tag.into_buf(), |b| settings.display_two_rows = b)
-            } else if tag.name() == b"DropDecimals" {
-                parse_bool(reader, tag.into_buf(), |b| settings.drop_decimals = b)
-            } else {
-                end_tag(reader, tag.into_buf())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    if !override_label {
-        settings.label_color = None;
-    }
-    settings.background = background_builder.build();
-
-    Ok(())
-}
-
-fn detailed_timer_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut detailed_timer::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let mut settings = component.settings().clone();
-    let mut background_builder = GradientBuilder::new();
-    let mut timer_override_color = false;
-    let (mut total_height, mut segment_timer_ratio) = (65u32, 0.4);
-
-    parse_children(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if tag.name() == b"Height" {
-                text_parsed(reader, tag.into_buf(), |v| total_height = v)
-            } else if tag.name() == b"SegmentTimerSizeRatio" {
-                text_parsed(reader, tag.into_buf(), |v: u32| {
-                    segment_timer_ratio = v as f32 / 100.0
-                })
-            } else if tag.name() == b"TimerShowGradient" {
-                parse_bool(reader, tag.into_buf(), |b| settings.timer.show_gradient = b)
-            } else if tag.name() == b"OverrideTimerColors" {
-                // Version >= 1.3
-                parse_bool(reader, tag.into_buf(), |b| timer_override_color = b)
-            } else if tag.name() == b"TimerUseSplitColors" {
-                // Version < 1.3
-                parse_bool(reader, tag.into_buf(), |b| timer_override_color = !b)
-            } else if tag.name() == b"SegmentTimerShowGradient" {
-                parse_bool(reader, tag.into_buf(), |b| {
-                    settings.segment_timer.show_gradient = b
-                })
-            } else if tag.name() == b"TimerFormat" {
-                // Version >= 1.5
-                timer_format(reader, tag.into_buf(), |d, a| {
-                    settings.timer.digits_format = d;
-                    settings.timer.accuracy = a;
-                })
-            } else if tag.name() == b"SegmentTimerFormat" {
-                // Version >= 1.5
-                timer_format(reader, tag.into_buf(), |d, a| {
-                    settings.segment_timer.digits_format = d;
-                    settings.segment_timer.accuracy = a;
-                })
-            } else if tag.name() == b"TimerAccuracy" {
-                // Version < 1.5
-                settings.timer.digits_format = DigitsFormat::SingleDigitSeconds;
-                accuracy(reader, tag.into_buf(), |v| settings.timer.accuracy = v)
-            } else if tag.name() == b"SegmentTimerAccuracy" {
-                // Version < 1.5
-                settings.segment_timer.digits_format = DigitsFormat::SingleDigitSeconds;
-                accuracy(reader, tag.into_buf(), |v| {
-                    settings.segment_timer.accuracy = v
-                })
-            } else if tag.name() == b"TimerColor" {
-                color(reader, tag.into_buf(), |v| {
-                    settings.timer.color_override = Some(v)
-                })
-            } else if tag.name() == b"DisplayIcon" {
-                parse_bool(reader, tag.into_buf(), |b| settings.display_icon = b)
-            } else if tag.name() == b"ShowSplitName" {
-                parse_bool(reader, tag.into_buf(), |b| settings.show_segment_name = b)
-            } else if tag.name() == b"Comparison" {
-                comparison_override(reader, tag.into_buf(), |v| settings.comparison1 = v)
-            } else if tag.name() == b"Comparison2" {
-                comparison_override(reader, tag.into_buf(), |v| settings.comparison2 = v)
-            } else if tag.name() == b"HideComparison" {
-                parse_bool(reader, tag.into_buf(), |b| {
-                    settings.hide_second_comparison = b
-                })
-            } else if tag.name() == b"TimingMethod" {
-                timing_method_override(reader, tag.into_buf(), |v| settings.timer.timing_method = v)
-            } else {
-                // FIXME:
-                // Width
-                // SegmentTimesAccuracy
-                // SegmentTimerColor
-                // SegmentLabelsColor
-                // SegmentTimesColor
-                // SegmentLabelsFont
-                // SegmentTimesFont
-                // SplitNameFont
-                // IconSize
-                // SplitNameColor
-                // DecimalsSize
-                // SegmentTimerDecimalsSize
-                end_tag(reader, tag.into_buf())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    if !timer_override_color {
-        // FIXME: This isn't actually exposed in the Detailed Timer's settings.
-        settings.timer.color_override = None;
-    }
-    settings.background = background_builder.build();
-
-    settings.segment_timer.height = (total_height as f32 * segment_timer_ratio) as u32;
-    settings.timer.height = total_height - settings.segment_timer.height;
-
-    component.set_settings(settings);
-
-    Ok(())
-}
-
-fn graph_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut graph::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-
-    parse_children(reader, buf, |reader, tag| {
-        if tag.name() == b"Height" {
-            text_parsed(reader, tag.into_buf(), |v| settings.height = v)
-        } else if tag.name() == b"BehindGraphColor" {
-            color(reader, tag.into_buf(), |c| {
-                settings.behind_background_color = c
-            })
-        } else if tag.name() == b"AheadGraphColor" {
-            color(reader, tag.into_buf(), |c| {
-                settings.ahead_background_color = c
-            })
-        } else if tag.name() == b"GridlinesColor" {
-            color(reader, tag.into_buf(), |c| {
-                println!("{:?} <- {:?}", settings.grid_lines_color, c);
-                settings.grid_lines_color = c
-            })
-        } else if tag.name() == b"PartialFillColorAhead" {
-            // Version >= 1.2
-            color(reader, tag.into_buf(), |c| settings.partial_fill_color = c)
-        } else if tag.name() == b"CompleteFillColorAhead" {
-            // Version >= 1.2
-            color(reader, tag.into_buf(), |c| settings.complete_fill_color = c)
-        } else if tag.name() == b"PartialFillColor" {
-            // Version < 1.2
-            color(reader, tag.into_buf(), |c| settings.partial_fill_color = c)
-        } else if tag.name() == b"CompleteFillColor" {
-            // Version < 1.2
-            color(reader, tag.into_buf(), |c| settings.complete_fill_color = c)
-        } else if tag.name() == b"GraphColor" {
-            color(reader, tag.into_buf(), |c| settings.graph_lines_color = c)
-        } else if tag.name() == b"LiveGraph" {
-            parse_bool(reader, tag.into_buf(), |b| settings.live_graph = b)
-        } else if tag.name() == b"FlipGraph" {
-            parse_bool(reader, tag.into_buf(), |b| settings.flip_graph = b)
-        } else if tag.name() == b"Comparison" {
-            comparison_override(reader, tag.into_buf(), |v| settings.comparison_override = v)
-        } else if tag.name() == b"ShowBestSegments" {
-            parse_bool(reader, tag.into_buf(), |b| settings.show_best_segments = b)
-        } else {
-            // FIXME:
-            // Width
-            // PartialFillColorBehind // Version >= 1.2
-            // CompleteFillColorBehind // Version >= 1.2
-            // ShadowsColor
-            // GraphLinesColor (separators, not our graph_lines_color)
-            // GraphGoldColor
-            end_tag(reader, tag.into_buf())
-        }
-    })
-}
-
-fn possible_time_save_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut possible_time_save::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-    let mut background_builder = GradientBuilder::new();
-    let (mut override_label, mut override_value) = (false, false);
-
-    parse_children(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if tag.name() == b"TextColor" {
-                color(reader, tag.into_buf(), |c| settings.label_color = Some(c))
-            } else if tag.name() == b"OverrideTextColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_label = b)
-            } else if tag.name() == b"TimeColor" {
-                color(reader, tag.into_buf(), |c| settings.value_color = Some(c))
-            } else if tag.name() == b"OverrideTimeColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_value = b)
-            } else if tag.name() == b"Accuracy" {
-                accuracy(reader, tag.into_buf(), |v| settings.accuracy = v)
-            } else if tag.name() == b"Comparison" {
-                comparison_override(reader, tag.into_buf(), |v| settings.comparison_override = v)
-            } else if tag.name() == b"Display2Rows" {
-                parse_bool(reader, tag.into_buf(), |b| settings.display_two_rows = b)
-            } else if tag.name() == b"TotalTimeSave" {
-                parse_bool(reader, tag.into_buf(), |b| {
-                    settings.total_possible_time_save = b
-                })
-            } else {
-                end_tag(reader, tag.into_buf())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    if !override_label {
-        settings.label_color = None;
-    }
-    if !override_value {
-        settings.value_color = None;
-    }
-    settings.background = background_builder.build();
-
-    Ok(())
-}
-
-fn previous_segment_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut previous_segment::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-    let mut background_builder = GradientBuilder::new();
-    let mut override_label = false;
-
-    parse_children(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if tag.name() == b"TextColor" {
-                color(reader, tag.into_buf(), |c| settings.label_color = Some(c))
-            } else if tag.name() == b"OverrideTextColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_label = b)
-            } else if tag.name() == b"DeltaAccuracy" {
-                accuracy(reader, tag.into_buf(), |v| settings.accuracy = v)
-            } else if tag.name() == b"DropDecimals" {
-                parse_bool(reader, tag.into_buf(), |b| settings.drop_decimals = b)
-            } else if tag.name() == b"Comparison" {
-                comparison_override(reader, tag.into_buf(), |v| settings.comparison_override = v)
-            } else if tag.name() == b"Display2Rows" {
-                parse_bool(reader, tag.into_buf(), |b| settings.display_two_rows = b)
-            } else if tag.name() == b"ShowPossibleTimeSave" {
-                parse_bool(reader, tag.into_buf(), |b| {
-                    settings.show_possible_time_save = b
-                })
-            } else {
-                // FIXME:
-                // TimeSaveAccuracy
-                end_tag(reader, tag.into_buf())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    if !override_label {
-        settings.label_color = None;
-    }
-    settings.background = background_builder.build();
-
-    Ok(())
-}
-
-fn splits_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut splits::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-    let mut split_gradient_builder = GradientBuilder::<GradientKind>::with_tags(
-        b"CurrentSplitTopColor",
-        b"CurrentSplitBottomColor",
-        b"CurrentSplitGradient",
-    );
-    let mut background_builder = GradientBuilder::<ListGradientKind>::new_gradient_type();
-
-    parse_children(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if let Some(tag) = split_gradient_builder.parse_background(reader, tag)? {
-                if tag.name() == b"VisualSplitCount" {
-                    text_parsed(reader, tag.into_buf(), |v| settings.visual_split_count = v)
-                } else if tag.name() == b"SplitPreviewCount" {
-                    text_parsed(reader, tag.into_buf(), |v| settings.split_preview_count = v)
-                } else if tag.name() == b"ShowThinSeparators" {
-                    parse_bool(reader, tag.into_buf(), |b| {
-                        settings.show_thin_separators = b
-                    })
-                } else if tag.name() == b"AlwaysShowLastSplit" {
-                    parse_bool(reader, tag.into_buf(), |b| {
-                        settings.always_show_last_split = b
-                    })
-                } else if tag.name() == b"SplitPreviewCount" {
-                    text_parsed(reader, tag.into_buf(), |v| settings.split_preview_count = v)
-                } else if tag.name() == b"ShowBlankSplits" {
-                    parse_bool(reader, tag.into_buf(), |b| {
-                        settings.fill_with_blank_space = b
-                    })
-                } else if tag.name() == b"SeparatorLastSplit" {
-                    parse_bool(reader, tag.into_buf(), |b| {
-                        settings.separator_last_split = b
-                    })
-                } else if tag.name() == b"Display2Rows" {
-                    parse_bool(reader, tag.into_buf(), |b| settings.display_two_rows = b)
-                } else if tag.name() == b"ShowColumnLabels" {
-                    parse_bool(reader, tag.into_buf(), |b| settings.show_column_labels = b)
-                } else if tag.name() == b"Columns" {
-                    // Version >= 1.5
-                    settings.columns.clear();
-
-                    parse_children(reader, tag.into_buf(), |reader, tag| {
-                        let mut column = splits::ColumnSettings::default();
-                        parse_children(reader, tag.into_buf(), |reader, tag| {
-                            if tag.name() == b"Name" {
-                                text(reader, tag.into_buf(), |v| column.name = v.into_owned())
-                            } else if tag.name() == b"Comparison" {
-                                comparison_override(reader, tag.into_buf(), |v| {
-                                    column.comparison_override = v
-                                })
-                            } else if tag.name() == b"TimingMethod" {
-                                timing_method_override(reader, tag.into_buf(), |v| {
-                                    column.timing_method = v
-                                })
-                            } else if tag.name() == b"Type" {
-                                text_err(reader, tag.into_buf(), |v| {
-                                    use self::splits::{
-                                        ColumnStartWith, ColumnUpdateTrigger, ColumnUpdateWith,
-                                    };
-                                    let (start_with, update_with, update_trigger) = match &*v {
-                                        "Delta" => (
-                                            ColumnStartWith::Empty,
-                                            ColumnUpdateWith::Delta,
-                                            ColumnUpdateTrigger::Contextual,
-                                        ),
-                                        "SplitTime" => (
-                                            ColumnStartWith::ComparisonTime,
-                                            ColumnUpdateWith::SplitTime,
-                                            ColumnUpdateTrigger::OnEndingSegment,
-                                        ),
-                                        "DeltaorSplitTime" => (
-                                            ColumnStartWith::ComparisonTime,
-                                            ColumnUpdateWith::Delta, // TODO: With Fallback
-                                            ColumnUpdateTrigger::Contextual,
-                                        ),
-                                        "SegmentDelta" => (
-                                            ColumnStartWith::Empty,
-                                            ColumnUpdateWith::SegmentDelta,
-                                            ColumnUpdateTrigger::Contextual,
-                                        ),
-                                        "SegmentTime" => (
-                                            ColumnStartWith::ComparisonSegmentTime,
-                                            ColumnUpdateWith::SegmentTime,
-                                            ColumnUpdateTrigger::OnEndingSegment,
-                                        ),
-                                        "SegmentDeltaorSegmentTime" => (
-                                            ColumnStartWith::ComparisonSegmentTime,
-                                            ColumnUpdateWith::SegmentDelta, // TODO: With Fallback
-                                            ColumnUpdateTrigger::Contextual,
-                                        ),
-                                        _ => return Err(Error::ColumnType),
-                                    };
-                                    column.start_with = start_with;
-                                    column.update_with = update_with;
-                                    column.update_trigger = update_trigger;
-                                    Ok(())
-                                })
-                            } else {
-                                end_tag(reader, tag.into_buf())
-                            }
-                        })?;
-                        settings.columns.insert(0, column);
-                        Ok(())
-                    })
-                } else if tag.name() == b"Comparison" {
-                    // Version < 1.5
-                    comparison_override(reader, tag.into_buf(), |v| {
-                        for column in &mut settings.columns {
-                            column.comparison_override = v.clone();
-                        }
-                    })
-                } else if tag.name() == b"ShowSplitTimes" {
-                    // Version < 1.5
-                    use self::splits::{
-                        ColumnSettings, ColumnStartWith, ColumnUpdateTrigger, ColumnUpdateWith,
-                    };
-                    parse_bool(reader, tag.into_buf(), |b| {
-                        if !b {
-                            let comparison_override =
-                                settings.columns.pop().and_then(|c| c.comparison_override);
-                            settings.columns.clear();
-                            // TODO: Write a test that verifies the order of the
-                            // two columns (The assumption here is that it's
-                            // right to left).
-                            settings.columns.push(ColumnSettings {
-                                name: String::from("Time"),
-                                start_with: ColumnStartWith::ComparisonTime,
-                                update_with: ColumnUpdateWith::SplitTime,
-                                update_trigger: ColumnUpdateTrigger::OnEndingSegment,
-                                comparison_override: comparison_override.clone(),
-                                timing_method: None,
-                            });
-                            settings.columns.push(ColumnSettings {
-                                name: String::from("+/−"),
-                                start_with: ColumnStartWith::Empty,
-                                update_with: ColumnUpdateWith::Delta,
-                                update_trigger: ColumnUpdateTrigger::Contextual,
-                                comparison_override: comparison_override,
-                                timing_method: None,
-                            });
-                        }
-                    })
-                } else {
-                    // FIXME:
-                    // DisplayIcons
-                    // SplitWidth
-                    // SplitTimesAccuracy
-                    // AutomaticAbbreviations
-                    // BeforeNamesColor // Version >= 1.3
-                    // CurrentNamesColor // Version >= 1.3
-                    // AfterNamesColor // Version >= 1.3
-                    // OverrideTextColor // Version >= 1.3
-                    // SplitNamesColor // Version >= 1.2 && Version < 1.3
-                    // UseTextColor // Version < 1.3
-                    // BeforeTimesColor
-                    // CurrentTimesColor
-                    // AfterTimesColor
-                    // OverrideTimesColor
-                    // LockLastSplit
-                    // IconSize
-                    // IconShadows
-                    // SplitHeight
-                    // DeltasAccuracy
-                    // DropDecimals
-                    // OverrideDeltasColor
-                    // DeltasColor
-                    // LabelsColor
-                    end_tag(reader, tag.into_buf())
-                }
-            } else {
-                Ok(())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    settings.current_split_gradient = split_gradient_builder.build();
-    settings.background = background_builder.build();
-
-    Ok(())
-}
-
-fn sum_of_best_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut sum_of_best::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-    let mut background_builder = GradientBuilder::new();
-    let (mut override_label, mut override_value) = (false, false);
-
-    parse_children(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if tag.name() == b"TextColor" {
-                color(reader, tag.into_buf(), |c| settings.label_color = Some(c))
-            } else if tag.name() == b"OverrideTextColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_label = b)
-            } else if tag.name() == b"TimeColor" {
-                color(reader, tag.into_buf(), |c| settings.value_color = Some(c))
-            } else if tag.name() == b"OverrideTimeColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_value = b)
-            } else if tag.name() == b"Accuracy" {
-                accuracy(reader, tag.into_buf(), |v| settings.accuracy = v)
-            } else if tag.name() == b"Display2Rows" {
-                parse_bool(reader, tag.into_buf(), |b| settings.display_two_rows = b)
-            } else {
-                end_tag(reader, tag.into_buf())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    if !override_label {
-        settings.label_color = None;
-    }
-    if !override_value {
-        settings.value_color = None;
-    }
-    settings.background = background_builder.build();
-
-    Ok(())
-}
-
-fn text_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut text::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-    let mut background_builder = GradientBuilder::new();
-    let (mut override_label, mut override_value) = (false, false);
-    let (mut left_center, mut right) = (String::new(), String::new());
-
-    parse_children(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if tag.name() == b"TextColor" {
-                color(reader, tag.into_buf(), |c| {
-                    settings.left_center_color = Some(c)
-                })
-            } else if tag.name() == b"OverrideTextColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_label = b)
-            } else if tag.name() == b"TimeColor" {
-                color(reader, tag.into_buf(), |c| settings.right_color = Some(c))
-            } else if tag.name() == b"OverrideTimeColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_value = b)
-            } else if tag.name() == b"Text1" {
-                text(reader, tag.into_buf(), |v| left_center = v.into_owned())
-            } else if tag.name() == b"Text2" {
-                text(reader, tag.into_buf(), |v| right = v.into_owned())
-            } else if tag.name() == b"Display2Rows" {
-                parse_bool(reader, tag.into_buf(), |b| settings.display_two_rows = b)
-            } else {
-                // FIXME:
-                // Font1
-                // Font2
-                // OverrideFont1
-                // OverrideFont2
-                end_tag(reader, tag.into_buf())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    if !override_label {
-        settings.left_center_color = None;
-    }
-    if !override_value {
-        settings.right_color = None;
-    }
-    settings.text = match (left_center.is_empty(), right.is_empty()) {
-        (false, false) => text::Text::Split(left_center, right),
-        (false, true) => text::Text::Center(left_center),
-        _ => text::Text::Center(right),
-    };
-    settings.background = background_builder.build();
-
-    Ok(())
-}
-
-fn timer_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut timer::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-    let mut background_builder = GradientBuilder::new();
-    let mut override_color = false;
-
-    parse_children(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if tag.name() == b"TimerHeight" {
-                text_parsed(reader, tag.into_buf(), |v| settings.height = v)
-            } else if tag.name() == b"TimerFormat" {
-                // Version >= 1.5
-                timer_format(reader, tag.into_buf(), |d, a| {
-                    settings.digits_format = d;
-                    settings.accuracy = a;
-                })
-            } else if tag.name() == b"TimerAccuracy" {
-                // Version >= 1.2 && Version < 1.5
-                accuracy(reader, tag.into_buf(), |v| settings.accuracy = v)
-            } else if tag.name() == b"OverrideSplitColors" {
-                // Version >= 1.3
-                parse_bool(reader, tag.into_buf(), |b| override_color = b)
-            } else if tag.name() == b"UseSplitColors" {
-                // Version < 1.3
-                parse_bool(reader, tag.into_buf(), |b| override_color = !b)
-            } else if tag.name() == b"ShowGradient" {
-                parse_bool(reader, tag.into_buf(), |b| settings.show_gradient = b)
-            } else if tag.name() == b"TimerColor" {
-                color(reader, tag.into_buf(), |c| {
-                    settings.color_override = Some(c)
-                })
-            } else if tag.name() == b"TimingMethod" {
-                timing_method_override(reader, tag.into_buf(), |v| settings.timing_method = v)
-            } else {
-                // FIXME:
-                // TimerWidth
-                // CenterTimer
-                // DecimalsSize
-                end_tag(reader, tag.into_buf())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    if !override_color {
-        settings.color_override = None;
-    }
-    settings.background = background_builder.build();
-
-    Ok(())
-}
-
-fn title_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut title::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-    let mut background_builder = GradientBuilder::new();
-    let mut override_title_color = false;
-
-    parse_children(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if tag.name() == b"ShowGameName" {
-                parse_bool(reader, tag.into_buf(), |b| settings.show_game_name = b)
-            } else if tag.name() == b"ShowCategoryName" {
-                parse_bool(reader, tag.into_buf(), |b| settings.show_category_name = b)
-            } else if tag.name() == b"ShowAttemptCount" {
-                parse_bool(reader, tag.into_buf(), |b| settings.show_attempt_count = b)
-            } else if tag.name() == b"ShowFinishedRunsCount" {
-                parse_bool(reader, tag.into_buf(), |b| {
-                    settings.show_finished_runs_count = b
-                })
-            } else if tag.name() == b"OverrideTitleColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_title_color = b)
-            } else if tag.name() == b"TextAlignment" {
-                // Version >= 1.7.3
-                text_as_escaped_bytes_err(reader, tag.into_buf(), |v| {
-                    settings.text_alignment = match &*v {
-                        b"0" => Alignment::Auto,
-                        b"1" => Alignment::Left,
-                        b"2" => Alignment::Center,
-                        _ => return Err(Error::Alignment),
-                    };
-                    Ok(())
-                })
-            } else if tag.name() == b"CenterTitle" {
-                // Version >= 1.3 && Version < 1.7.3
-                parse_bool(reader, tag.into_buf(), |b| {
-                    settings.text_alignment = if b {
-                        Alignment::Center
-                    } else {
-                        Alignment::Auto
-                    }
-                })
-            } else if tag.name() == b"SingleLine" {
-                parse_bool(reader, tag.into_buf(), |b| {
-                    settings.display_as_single_line = b
-                })
-            } else if tag.name() == b"TitleColor" {
-                color(reader, tag.into_buf(), |c| settings.text_color = Some(c))
-            } else if tag.name() == b"DisplayGameIcon" {
-                parse_bool(reader, tag.into_buf(), |b| settings.display_game_icon = b)
-            } else if tag.name() == b"ShowRegion" {
-                parse_bool(reader, tag.into_buf(), |b| settings.show_region = b)
-            } else if tag.name() == b"ShowPlatform" {
-                parse_bool(reader, tag.into_buf(), |b| settings.show_platform = b)
-            } else if tag.name() == b"ShowVariables" {
-                parse_bool(reader, tag.into_buf(), |b| settings.show_variables = b)
-            } else {
-                // FIXME:
-                // OverrideTitleFont // Version >= 1.3
-                // TitleFont // Version >= 1.2
-                // UseLayoutSettingsFont // Version >= 1.2 && Version < 1.3
-                end_tag(reader, tag.into_buf())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    if !override_title_color {
-        settings.text_color = None;
-    }
-    settings.background = background_builder.build();
-
-    Ok(())
-}
-
-fn total_playtime_settings<R>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    component: &mut total_playtime::Component,
-) -> Result<()>
-where
-    R: BufRead,
-{
-    let settings = component.settings_mut();
-    let mut background_builder = GradientBuilder::new();
-    let (mut override_label, mut override_value) = (false, false);
-
-    parse_children(reader, buf, |reader, tag| {
-        if let Some(tag) = background_builder.parse_background(reader, tag)? {
-            if tag.name() == b"TextColor" {
-                color(reader, tag.into_buf(), |c| settings.label_color = Some(c))
-            } else if tag.name() == b"OverrideTextColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_label = b)
-            } else if tag.name() == b"TimeColor" {
-                color(reader, tag.into_buf(), |c| settings.value_color = Some(c))
-            } else if tag.name() == b"OverrideTimeColor" {
-                parse_bool(reader, tag.into_buf(), |b| override_value = b)
-            } else if tag.name() == b"Display2Rows" {
-                parse_bool(reader, tag.into_buf(), |b| settings.display_two_rows = b)
-            } else if tag.name() == b"ShowTotalHours" {
-                parse_bool(reader, tag.into_buf(), |b| settings.show_days = !b)
-            } else {
-                end_tag(reader, tag.into_buf())
-            }
-        } else {
-            Ok(())
-        }
-    })?;
-
-    if !override_label {
-        settings.label_color = None;
-    }
-    if !override_value {
-        settings.value_color = None;
-    }
-    settings.background = background_builder.build();
-
-    Ok(())
-}
-
 fn component<R, F>(reader: &mut Reader<R>, buf: &mut Vec<u8>, f: F) -> Result<()>
 where
     R: BufRead,
@@ -1270,30 +368,30 @@ where
             // Otherwise we need to cache the settings and load them later.
             if let Some(component) = &mut component {
                 match component {
-                    Component::BlankSpace(c) => blank_space_settings(reader, tag.into_buf(), c),
+                    Component::BlankSpace(c) => blank_space::settings(reader, tag.into_buf(), c),
                     Component::CurrentComparison(c) => {
-                        current_comparison_settings(reader, tag.into_buf(), c)
+                        current_comparison::settings(reader, tag.into_buf(), c)
                     }
-                    Component::CurrentPace(c) => current_pace_settings(reader, tag.into_buf(), c),
-                    Component::Delta(c) => delta_settings(reader, tag.into_buf(), c),
+                    Component::CurrentPace(c) => current_pace::settings(reader, tag.into_buf(), c),
+                    Component::Delta(c) => delta::settings(reader, tag.into_buf(), c),
                     Component::DetailedTimer(c) => {
-                        detailed_timer_settings(reader, tag.into_buf(), c)
+                        detailed_timer::settings(reader, tag.into_buf(), c)
                     }
-                    Component::Graph(c) => graph_settings(reader, tag.into_buf(), c),
+                    Component::Graph(c) => graph::settings(reader, tag.into_buf(), c),
                     Component::PossibleTimeSave(c) => {
-                        possible_time_save_settings(reader, tag.into_buf(), c)
+                        possible_time_save::settings(reader, tag.into_buf(), c)
                     }
                     Component::PreviousSegment(c) => {
-                        previous_segment_settings(reader, tag.into_buf(), c)
+                        previous_segment::settings(reader, tag.into_buf(), c)
                     }
                     Component::Separator(_) => end_tag(reader, tag.into_buf()),
-                    Component::Splits(c) => splits_settings(reader, tag.into_buf(), c),
-                    Component::SumOfBest(c) => sum_of_best_settings(reader, tag.into_buf(), c),
-                    Component::Text(c) => text_settings(reader, tag.into_buf(), c),
-                    Component::Timer(c) => timer_settings(reader, tag.into_buf(), c),
-                    Component::Title(c) => title_settings(reader, tag.into_buf(), c),
+                    Component::Splits(c) => splits::settings(reader, tag.into_buf(), c),
+                    Component::SumOfBest(c) => sum_of_best::settings(reader, tag.into_buf(), c),
+                    Component::Text(c) => text::settings(reader, tag.into_buf(), c),
+                    Component::Timer(c) => timer::settings(reader, tag.into_buf(), c),
+                    Component::Title(c) => title::settings(reader, tag.into_buf(), c),
                     Component::TotalPlaytime(c) => {
-                        total_playtime_settings(reader, tag.into_buf(), c)
+                        total_playtime::settings(reader, tag.into_buf(), c)
                     }
                 }
             } else {
